+++ templates/partials/corporation/corporation-logo.php
@@ -19,7 +19,7 @@
 
 use \WordPress\Plugins\EveOnlineIntelTool\Libs\Helper\ImageHelper;
 
-$imageCorporation = ImageHelper::getInstance()->getImageServerUrl('corporation') . $data['corporationID'] . '_32.png';
+$corporationLogo = ImageHelper::getInstance()->getImageServerUrl('corporation') . $data['corporationID'] . '_32.png';
 
 $size = 32;
 if(isset($data['size'])) {
@@ -27,4 +27,4 @@ if(isset($data['size'])) {
 }
 ?>
 
-<span class="eve-intel-corporation-logo-wrapper"><img class="eve-image" data-eveid="<?php echo $data['corporationID']; ?>" src="<?php echo $imageCorporation; ?>" alt="<?php echo $data['corporationName']; ?>" title="<?php echo $data['corporationName']; ?>" width="<?php echo $size; ?>" heigh="<?php echo $size; ?>"></span>
+<span class="eve-intel-corporation-logo-wrapper"><img class="eve-image" data-eveid="<?php echo $data['corporationID']; ?>" src="<?php echo $corporationLogo; ?>" alt="<?php echo $data['corporationName']; ?>" title="<?php echo $data['corporationName']; ?>" width="<?php echo $size; ?>" heigh="<?php echo $size; ?>"></span>