+++ templates/partials/pilot/pilot-avatar.php
@@ -19,7 +19,7 @@
 
 use \WordPress\Plugins\EveOnlineIntelTool\Libs\Helper\ImageHelper;
 
-$imagePilot = ImageHelper::getInstance()->getImageServerUrl('character') . $data['characterID'] . '_32.jpg';
+$characterPortrait = ImageHelper::getInstance()->getImageServerUrl('character') . $data['characterID'] . '_32.jpg';
 ?>
 
-<span class="eve-intel-pilot-avatar-wrapper"><img class="eve-image" data-eveid="<?php echo $data['characterID']; ?>" src="<?php echo $imagePilot; ?>" alt="<?php echo $data['characterName']; ?>" width="32" heigh="32"></span>
+<span class="eve-intel-pilot-avatar-wrapper"><img class="eve-image" data-eveid="<?php echo $data['characterID']; ?>" src="<?php echo $characterPortrait; ?>" alt="<?php echo $data['characterName']; ?>" width="32" heigh="32"></span>